+++ manifest.go
@@ -40,6 +40,7 @@ var (
 	errRootPruneContainsName   = errors.Errorf("%q should not include a name", "prune")
 	errInvalidRootPruneValue   = errors.New("root prune options must be omitted instead of being set to false")
 	errInvalidPruneProjectName = errors.Errorf("%q in %q must be a string", "name", "prune.project")
+	errNoName                  = errors.New("no name provided")
 )
 
 // Manifest holds manifest file data and implements gps.RootManifest.
@@ -133,13 +134,17 @@ func validateManifest(s string) ([]error, error) {
 				if valid {
 					// Iterate through each array of tables
 					for _, v := range rawProj {
+						ruleProvided := false
+						props := v.(map[string]interface{})
 						// Check the individual field's key to be valid
-						for key, value := range v.(map[string]interface{}) {
+						for key, value := range props {
 							// Check if the key is valid
 							switch key {
-							case "name", "branch", "version", "source":
-								// valid key
+							case "name":
+							case "branch", "version", "source":
+								ruleProvided = true
 							case "revision":
+								ruleProvided = true
 								if valueStr, ok := value.(string); ok {
 									if abbrevRevHash.MatchString(valueStr) {
 										warns = append(warns, fmt.Errorf("revision %q should not be in abbreviated form", valueStr))
@@ -154,6 +159,11 @@ func validateManifest(s string) ([]error, error) {
 								// unknown/invalid key
 								warns = append(warns, fmt.Errorf("invalid key %q in %q", key, prop))
 							}
+						}
+						if _, ok := props["name"]; !ok {
+							warns = append(warns, errNoName)
+						} else if !ruleProvided && prop == "constraint" {
+							warns = append(warns, fmt.Errorf("branch, version, revision, or source should be provided for %q", props["name"]))
 						}
 					}
 				}